requirements.txt
torch>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
wandb>=0.15.0
pyyaml>=6.0
stable-retro
gymnasium
